Move SecretGenerator out of Config
SecretGenerator is used to create web session secrets and doesn't really belong to config in any way.
Moving inside Application so it can still be mocked in tests.
