+++ core/chains/evm/config/mocks/chain_scoped_config.go
@@ -4074,29 +4074,6 @@ func (_m *ChainScopedConfig) SessionOptions() sessions.Options {
 	return r0
 }
 
-// SessionSecret provides a mock function with given fields:
-func (_m *ChainScopedConfig) SessionSecret() ([]byte, error) {
-	ret := _m.Called()
-
-	var r0 []byte
-	if rf, ok := ret.Get(0).(func() []byte); ok {
-		r0 = rf()
-	} else {
-		if ret.Get(0) != nil {
-			r0 = ret.Get(0).([]byte)
-		}
-	}
-
-	var r1 error
-	if rf, ok := ret.Get(1).(func() error); ok {
-		r1 = rf()
-	} else {
-		r1 = ret.Error(1)
-	}
-
-	return r0, r1
-}
-
 // SessionTimeout provides a mock function with given fields:
 func (_m *ChainScopedConfig) SessionTimeout() models.Duration {
 	ret := _m.Called()
