+++ core/services/chainlink/application.go
@@ -103,6 +103,8 @@ type Application interface {
 
 	// ID is unique to this particular application instance
 	ID() uuid.UUID
+
+	SecretGenerator() SecretGenerator
 }
 
 // ChainlinkApplication contains fields for the JobSubscriber, Scheduler,
@@ -132,6 +134,7 @@ type ChainlinkApplication struct {
 	logger                   logger.Logger
 	closeLogger              func() error
 	sqlxDB                   *sqlx.DB
+	secretGenerator          SecretGenerator
 
 	started     bool
 	startStopMu sync.Mutex
@@ -149,6 +152,7 @@ type ApplicationOpts struct {
 	Version                  string
 	RestrictedHTTPClient     *http.Client
 	UnrestrictedHTTPClient   *http.Client
+	SecretGenerator          SecretGenerator
 }
 
 // Chains holds a ChainSet for each type of chain.
@@ -444,6 +448,7 @@ func NewApplication(opts ApplicationOpts) (Application, error) {
 		Nurse:                    nurse,
 		logger:                   globalLogger,
 		closeLogger:              opts.CloseLogger,
+		secretGenerator:          opts.SecretGenerator,
 
 		sqlxDB: opts.SqlxDB,
 
@@ -610,6 +615,10 @@ func (app *ChainlinkApplication) TxmORM() txmgr.ORM {
 
 func (app *ChainlinkApplication) GetExternalInitiatorManager() webhook.ExternalInitiatorManager {
 	return app.ExternalInitiatorManager
+}
+
+func (app *ChainlinkApplication) SecretGenerator() SecretGenerator {
+	return app.secretGenerator
 }
 
 // WakeSessionReaper wakes up the reaper to do its reaping.
